+++ ricaun.RevitTest.Console/Revit/Utils/RevitTestUtils.cs
@@ -14,6 +14,7 @@ using System.Threading;
 
 namespace ricaun.RevitTest.Console.Revit.Utils
 {
+
     /// <summary>
     /// RevitTestUtils
     /// </summary>
@@ -23,9 +24,10 @@ namespace ricaun.RevitTest.Console.Revit.Utils
         private const int SleepMillisecondsDebuggerAttached = 1000;
 
         private const int TimeoutMinutesDefault = 10;
+        private const int MinimalRevitVersion = 2019; // Application is not compatible with Revit 2018 or lower
 
         /// <summary>
-        /// Get Test Full Names using RevitInstallation if needed (Revit +2021)
+        /// Get Test Full Names using RevitInstallation
         /// </summary>
         /// <param name="filePath"></param>
         /// <returns></returns>
@@ -101,7 +103,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
             double timeoutMinutes = 0,
             params string[] testFilters)
         {
-            int timeoutNotBusyCountMax = 10;
+            int timeoutNotBusyCountMax = EnvironmentVariable.TimeoutNotBusyMaxSeconds;
 
             if (timeoutMinutes <= 0)
                 timeoutMinutes = TimeoutMinutesDefault;
@@ -111,11 +113,22 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                 RevitUtils.TryGetRevitVersion(fileToTest, out revitVersionNumber);
             }
 
-            if (!RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersionNumber, out RevitInstallation revitInstallationNotExist))
+            if (!RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersionNumber, out RevitInstallation revitInstallationExist))
             {
                 var exceptionRevitNotExist = new Exception($"Installed Revit with version {revitVersionNumber} or greater not found.");
 
                 var failTests = TestEngine.Fail(fileToTest, exceptionRevitNotExist, testFilters);
+                actionOutput.Invoke(failTests.ToJson());
+
+                Thread.Sleep(SleepMillisecondsBeforeFinish);
+                return;
+            }
+
+            if (revitInstallationExist.Version < MinimalRevitVersion)
+            {
+                var exceptionRevitVersion = new Exception($"Revit version {revitInstallationExist.Version} is not compatible with the application, minimal Revit version {MinimalRevitVersion}.");
+                
+                var failTests = TestEngine.Fail(fileToTest, exceptionRevitVersion, testFilters);
                 actionOutput.Invoke(failTests.ToJson());
 
                 Thread.Sleep(SleepMillisecondsBeforeFinish);
@@ -154,7 +167,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                         if (revitInstallation.TryGetProcess(out Process process) == false || forceToOpenNewRevit)
                         {
                             var startRevitLanguageArgument = RevitLanguageUtils.GetArgument(forceLanguageToRevit);
-                            startRevitLanguageArgument = string.Join(" ", startRevitLanguageArgument, Environment.GetEnvironmentVariable("RICAUN_REVITTEST_CONSOLE_PROCESS_ARGUMENTS"));
+                            startRevitLanguageArgument = string.Join(" ", startRevitLanguageArgument, EnvironmentVariable.ProcessArguments);
                             Log.WriteLine($"{revitInstallation}: Start {startRevitLanguageArgument}");
                             process = revitInstallation.Start(startRevitLanguageArgument);
                         }
@@ -198,7 +211,7 @@ namespace ricaun.RevitTest.Console.Revit.Utils
                             }
                         };
 
-                        int timeoutSeconds = (int) timeoutMinutes * 60;
+                        int timeoutSeconds = (int)timeoutMinutes * 60;
                         for (int i = 0; i <= timeoutSeconds; i++)
                         {
                             Thread.Sleep(1000);